🔧 chore: update publish script in workflows
Included the option '--no-git-checks' in the 'pnpm publish' command to streamline the publishing process and skip git checks. Also, updated README.md with information on always using the latest version of the SDK, providing guidance for CDN usage, npm installation, and a note emphasizing the importance of using '@latest' to ensure getting the most recent version. Added details on automated publishing triggered by pushes to the 'main' branch, outlining the steps involved and the availability of the SDK via jsDelivr and UNPKG after each publish.